+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
  
  - Added an apply method on base classes. Now most resources can simply `apply` files. 
+ - added tenant DNS config command to retrieve configuration
  - Commands can have aliases now. For example `find` and `get` are aliases for the same command. Or `list` and `ls` for short. 
 
 ### Updates